requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Chemistry
rdkit>=2023.3.1

# Data & config
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0

# Protein handling
biopython>=1.81

# Docking (install separately if issues arise)
meeko>=0.5.0
vina>=1.2.5
